round_robin/requirements.txt
Django
djangorestframework
requests
psutil
docker
uvicorn==0.27.1 
httpx==0.24.1
locust
python-dotenv
django-ratelimit